Fix imported/w3c/web-platform-tests/css/css-box/inheritance.html for margin-trim. https://bugs.webkit.org/show_bug.cgi?id=249366 rdar://103383727
Reviewed by Tim Nguyen.

The test was using a value for the margin-trim property that is not
valid. The value that it was using, "all," was in an old version of the
spec but is not valid anymore. This would cause the test to fail in any
scenarios since the parser would not recognize it as a value for the
property. Replacing it with "block," should help resolve this issue.

* LayoutTests/imported/w3c/web-platform-tests/css/css-box/inheritance.html:

Canonical link: https://commits.webkit.org/257903@main