[web-animations] account for iterationComposite when interpolating custom properties https://bugs.webkit.org/show_bug.cgi?id=249394
Reviewed by Antti Koivisto.

In order to correctly handle iterationComposite when interpolating custom properties, we
update CSSPropertyAnimation::propertyRequiresBlendingForAccumulativeIteration() to take a
CSSPropertyAnimation::Property as a parameter such that we can accept a custom property.

In the implementation we can check for the custom property type to identify whether it
requires blending for iterationComposite.

In KeyframeEffect::setAnimatedPropertiesInStyle(), we can then remove the code specific
to standard CSS properties when dealing with iterationComposite.

* LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/animation/custom-property-animation-color.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/animation/custom-property-animation-length-expected.txt:
* Source/WebCore/animation/CSSPropertyAnimation.cpp:
(WebCore::CSSPropertyAnimation::blendProperty):
(WebCore::CSSPropertyAnimation::propertyRequiresBlendingForAccumulativeIteration):
* Source/WebCore/animation/CSSPropertyAnimation.h:
* Source/WebCore/animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::setAnimatedPropertiesInStyle):

Canonical link: https://commits.webkit.org/257925@main