[CSS-Typed-OM] Reification of primitive values of type CSS_INTEGER should construct CSSUnitValues https://bugs.webkit.org/show_bug.cgi?id=249412
Reviewed by Simon Fraser.

Reification of primitive values of type CSS_INTEGER should construct
CSSUnitValues (with type 'number'), not generic CSSStyleValues.

* LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/order-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/orphans-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/widows-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/z-index-expected.txt:
Rebaseline WPT tests now that more checks are passing.

* LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/grid-start-end-expected.txt:
While this looks like a regression, I am not convinced the test is correct. The
test sets the `grid-row-start` property to 3 and expects a generic
CSSStyleValue back when querying the StylePropertyMap. However, since 3 is an
integer, we return a CSSUnitValue. I'll investigate this separately. I suspect
this may have to do with the fact that the `grid-row-start` value may be stored
as a list internally when an identifier is provided in addition to the integer.

* Source/WebCore/css/typedom/CSSStyleValueFactory.cpp:
(WebCore::CSSStyleValueFactory::reifyValue):

Canonical link: https://commits.webkit.org/257968@main