[GStreamer] Avoid crashes on caps generated from user input https://bugs.webkit.org/show_bug.cgi?id=249466
Reviewed by Philippe Normand.

When checking if certain mime is supported, we create the caps from it and if caps creation fails because input is
bogus, we can get a controlled crash later. With this patch, if we can't create the caps, we bail out more gracefully.

* Source/WebCore/platform/graphics/gstreamer/GStreamerRegistryScanner.cpp:
(WebCore::GStreamerRegistryScanner::ElementFactories::hasElementForMediaType const):

Canonical link: https://commits.webkit.org/257983@main