Use the fullscreen preferred size when reporting screen dimensions https://bugs.webkit.org/show_bug.cgi?id=249297 <rdar://99403187>
Reviewed by Megan Gardner.

When we have a preferred size for fullscreen windows, use it when
reporting the screen dimensions to the page.

* Source/WebKit/UIProcess/WebPageProxy.h:
Expose `overrideScreenSize`.

* Source/WebKit/UIProcess/ios/WebPageProxyIOS.mm:
(WebKit::WebPageProxy::overrideScreenSize):
Read the fullscreen preferred size and override the screen size with it.

* Source/WebKit/UIProcess/ios/fullscreen/WKFullScreenWindowControllerIOS.mm:
(-[WKFullScreenWindowController enterFullScreen:]):
Use the (modified) page's screen size as target size for the fullscreen
window.

Canonical link: https://commits.webkit.org/258005@main