Fix build for "[Xcode] IDECustomBuildProductsPaths starting with "~" are ignored when determining the build location" https://bugs.webkit.org/show_bug.cgi?id=249442
Unreviewed build fix.

That change failed to consider when there is _no_
IDECustomBuildProductsPaths, and someone is building from a directory
_other than_ the WebKit repo root. We check the
IDEApplicationwideBuildSettings defaults key, which may be an empty
dictionary "{ }", and were treating that as a path, rather than falling
back to using the repo's root directory to determine the location.

* Tools/Scripts/webkitdirs.pm:
(determineBaseProductDir):