Add webpushd iOS sandbox https://bugs.webkit.org/show_bug.cgi?id=249436 rdar://103410333
Reviewed by Brady Eidson.

This adds a sandbox profile for webpushd on iOS. It allows standard system operations, reading and
writing to the PushDatabase, interacting with apsd, and interacting with LaunchServices.

For now, the sandbox defaults to allowing all operations with symbolication, but we'll change the
default to deny once we're sure these rules are sufficient.

* Source/WebKit/DerivedSources-input.xcfilelist:
* Source/WebKit/Resources/SandboxProfiles/ios/com.apple.WebKit.webpushd.sb.in: Added.
* Source/WebKit/Scripts/process-entitlements.sh:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/258033@main