css rendering bug : fixed-position-element and 'left'
css rendering bug : fixed-position-element and 'left'
https://bugs.webkit.org/show_bug.cgi?id=122878

Reviewed by NOBODY (OOPS!).

This patch is to align Webkit with Gecko / Firefox and Blink / Chromium.

Merge - https://chromium.googlesource.com/chromium/blink/+/771dc814b28496295ded1ac67346d774f4ebd608

Fixed position element was rendered as normal flow irrespective of it's relative-position-parents.

If fixed position element does not have left and/or right position then
it should render as normal flow with respect to it's parent.
So fixed position element is shifted with the value of relative position
of it's parent.

* Source/WebCore/rendering/RenderBox.cpp:
(computeInlineStaticDistance): Update logic to account for left/right
position in reference to relative position
* LayoutTests/fast/css/css-properties-relative-as-parent-fixed.html: Add Test Case