Images are not loaded in element application https://bugs.webkit.org/show_bug.cgi?id=249213
Reviewed by Xabier Rodriguez-Calvar.

Element application uses the fetch api to download images and convert
them to a blob that is then requested. The problem is that the blob
request is failing due to invalid HTTP header in response. The invalid
header is Content-Security-Policy that is generated by the blob loader
using the security context policy. The document policy comes from
http-equiv meta and contains newlines, which are allowed there, but
invalid for an HTTP header value.

Test: http/tests/security/contentSecurityPolicy/blob-url-invalid-http-header.html

* LayoutTests/http/tests/security/contentSecurityPolicy/blob-url-invalid-http-header-expected.txt: Added.
* LayoutTests/http/tests/security/contentSecurityPolicy/blob-url-invalid-http-header.html: Added.
* Source/WebCore/page/csp/ContentSecurityPolicy.cpp:
(WebCore::ContentSecurityPolicy::responseHeaders const): Make the http header value valid if needed.

Canonical link: https://commits.webkit.org/258130@main