[Tools] built-product-archive shouldn't pass 'jsc' as port name for 'jsc-only' https://bugs.webkit.org/show_bug.cgi?id=249644
Reviewed by Michael Catanzaro.

'jsc' is not a valid port name, so built-product-archive ends calling
webkit-build-directory to determine the build directories with an unknown
port name, which returns the default directory layout.

This works by accident for the general use case, but it should be calling
webkit-build-directory script with the right port name.

Fix it to use and pass 'jsc-only' for the case of the JSCOnly port.

* Tools/CISupport/built-product-archive:

Canonical link: https://commits.webkit.org/258147@main