Call Document::willBeRemovedFromFrame before swapping a Frame for a RemoteFrame https://bugs.webkit.org/show_bug.cgi?id=249665 rdar://103565224
Reviewed by Simon Fraser.

This fixes crashes when using site isolation and loading cross-origin iframes
because the render tree is still attached to a Document which no longer has a
Frame before this change, but after this change the render tree is detached.
This makes investigation of bugs much easier and is the first step towards
hooking up drawing code for remote frames.

* Source/WebCore/dom/Document.h:
* Source/WebKit/WebProcess/WebPage/WebFrame.cpp:
(WebKit::WebFrame::didCommitLoadInAnotherProcess):

Canonical link: https://commits.webkit.org/258161@main