Prevent the WindowEventLoop tasks from running between trigger an rendering update and the end of that update https://bugs.webkit.org/show_bug.cgi?id=249684 rdar://103575958
Reviewed by NOBODY (OOPS!).

It's impossible to write non-flakey layout tests that involve async/await and requestAnimationFrame, because
their ordering depends on exactly when the Microtask that resumes the async function runs, and that's sensitive
to when the WindowEventLoop zero-delay timer fires.

Rendering updates are scheduled (generally at a 60Hz cadence) via display refresh callbacks, but each
one is then triggered from that callback via a CFRunLoopObserver in TiledCoreAnimationDrawingArea, and
a zero-delay timer in RemoteLayerTreeDrawingArea.

If the WindowEventLoop timer fires between scheduling the CFRunLoopObserver and it firing, then we get
a different ordering behavior, where an async function can resume with a different ordering relative to
requestAnimationFrame callbacks.

Fix by re-scheduling the WindowEventLoop timer if it fires between the `scheduleRenderingUpdateRunLoopObserver()`
and the `updateRenderingRunLoopCallback()` for TiledCoreAnimationDrawingArea.

RemoteLayerTreeDrawingArea uses timers, so this may not be as much of an issue there, but do the same thing for
consistency. Also, now that we're using RemoteLayerTreeDrawingArea on macOS with UI-side compositing, add the
call to `WindowEventLoop::breakToAllowRenderingUpdate()` there too.

* Source/WebCore/dom/WindowEventLoop.cpp:
(WebCore::WindowEventLoop::didReachTimeToRun):
(WebCore::WindowEventLoop::breakToAllowRenderingUpdate):
(WebCore::WindowEventLoop::didCompleteRenderingUpdate):
* Source/WebCore/dom/WindowEventLoop.h:
* Source/WebKit/WebProcess/WebPage/RemoteLayerTree/RemoteLayerTreeDrawingArea.h:
* Source/WebKit/WebProcess/WebPage/RemoteLayerTree/RemoteLayerTreeDrawingArea.mm:
(WebKit::RemoteLayerTreeDrawingArea::RemoteLayerTreeDrawingArea):
(WebKit::RemoteLayerTreeDrawingArea::triggerRenderingUpdate):
(WebKit::RemoteLayerTreeDrawingArea::updateRenderingTimerFired):
* Source/WebKit/WebProcess/WebPage/mac/TiledCoreAnimationDrawingArea.mm:
(WebKit::TiledCoreAnimationDrawingArea::updateRenderingRunLoopCallback):