Fix assertion failure when glyph ID 0 of a font that has a "VORG" table in vertical flow
Fix assertion failure when glyph ID 0 of a font that has a "VORG" table in vertical flow
https://bugs.webkit.org/show_bug.cgi?id=249317

Reviewed by NOBODY (OOPS!).

Merge - https://src.chromium.org/viewvc/blink?revision=195946&view=revision

In OpenType, glyph ID 0 is recommended for ".notdef"[1]. Rendering the
glyph of a font that has "VORG" table fails assertion in HashTable::lookup,
since 0 is not safeToCompareToEmptyOrDeleted.

It's a regular (non-RELEASE) ASSERT, but Mac and Adobe fonts usually
have this table, so some tests fail only on Mac.

Since it's a ".notdef" glyph, the fix is to use the default value.

[1] https://www.microsoft.com/typography/otspec/recom.htm

* Source/WebCore/platform/graphics/opentype/OpenTypeVerticalData.cpp:
(OpenTypeVerticalData::getVerticalTranslationForGlyphs): Add "if" condition for glyph
* LayoutTests/fast/writing-mode/vorg-glyph-zero-crash.html: Add Test Case
* LayoutTests/fast/writing-mode/vorg-glyph-zero-crash-expected.txt: Add Test Case Expectations