[WK2] Use Spans when encoding, decoding ArrayReferenceTuple objects https://bugs.webkit.org/show_bug.cgi?id=249727
Reviewed by NOBODY (OOPS!).

This adjusts the ArgumentCoder specialization for ArrayReferenceTuple
to use Span objects for encoding and decoding, along with some other
tweaks.

In encoding, Spans are adopted and used in a fold-expression over the
encoder's stream-insertion operator.

In decoding, the decoded size is retrieved and fitted into size_t. Then,
another fold-expression is used to detect any overflows when the size
value is multiplied by the byte-size for each type. Before, this was
done for each decoded tuple element separately. Data for each tuple
element is now decoded as a Span, corresponding to the new way of
encoding.

* Source/WebKit/Platform/IPC/ArgumentCoders.h: