[WPE] Temporarily disable ENBALE_COG by default https://bugs.webkit.org/show_bug.cgi?id=249791
Reviewed by NOBODY (OOPS!).

It's time to start making breaking changes for the WPE 2.0 API version,
but we cannot actually do this currently because our ENABLE_COG option
depends on an external checkout of cog that already depends on WPE 2.0.
Removing API that cog uses results in WPE port build failure. But we
also cannot plausibly change cog to use new APIs that do not yet exist
in WebKit because we cannot commit them because we have not yet updated
cog. Aaah, it's circular!

I see two workable options:

 1. WebKit must depend on side branches of cog, where cog is modified to
    account for WebKit API changes that have not yet landed in WebKit.
 2. Simply turn off ENABLE_COG for now, and reenable right before
    releasing 2.39.90, when the API stabilizes.

My preference is option 2 because that would be easier for WebKit
development, although it risks that we forget to keep cog in good shape.
We'll want to remember to turn cog on again in February when the API is
frozen.

* Source/cmake/OptionsWPE.cmake: