[WPE] Checkout a specific version of Cog instead of master https://bugs.webkit.org/show_bug.cgi?id=249034
Reviewed by NOBODY (OOPS!).

Igalia wants to start running performance tests, but performance
test results will be inconsistent if the version of cog may be different
from one day to the next. A specific commit should be used instead.

The downside of this change is we now have to remember to update the
Cog version to use in a rather unobvious place, Tools/PlatformWPE.cmake,
which is not something that we will remember to do on a regular basis.
We need a better way to handle this, such as adding it to the Cog
release process.

* Tools/PlatformWPE.cmake: