gdesktopappinfo.h not excluded on macOS https://bugs.webkit.org/show_bug.cgi?id=248727
Reviewed by Carlos Garcia Campos.

Although macOS has most of the gio-unix-2.0 functionality, this
particular header is an exception and just needs to be handled
specially.

MacPorts has a patch to sometimes build with GDesktopAppInfo enabled, so
it's best to just detect the presence of the header instead of trying to
enable this code on OS(UNIX_BUT_NOT_MACOS_EXCEPT_SOMETIMES).

* Source/WTF/wtf/PlatformHave.h:
* Source/WebKit/UIProcess/Notifications/glib/NotificationService.cpp:
(WebKit::applicationIcon):

Canonical link: https://commits.webkit.org/258260@main